refactor: align file and folder naming
