+++ dashboard/src/components/input/search-input/search-input.scss
